hitting enter while adding collection member, clicks correct Add
The default behavior on the collection member's section is that if you are in any field and hit enter, it will automatically submit whatever was the first submit_tag, which is the manager's Add button. If you are in the editor field or the depositor field and hit enter, it would automatically try to add a manager and it would give you the error: Editor can't be blank. In this fix I override the default behavior and tell it to click the correct add button.