MasterFile#working_file_path turned into unique path(s) persisted on the master file and passed to Matterhorn.
The working file path is only used when Settings.matterhorn.media_path is configured.  The code changes in prior commits did not take into account all of the ways ingest happens: web upload, web upload skip transcoding, web dropbox, web dropbox skip transcoding, batch ingest, batch ingest skip transcoding, and batch ingest pre-transcoded derivatives.  The approach to this commit was to write tests for each of these scenarios with both the Settings.matterhorn.media_path configuration set and not set then make all of the tests pass.  The underlying use case for the working file path is when Matterhorn is running on a different server than Avalon and they don't share the same filesystem (and filesystem paths) for both the Rails web upload directory and dropbox.  The working_file_path is a copy of the originally uploaded file to where Matterhorn has access and at a path shared by both servers.  For example, if Matterhorn is running on a different server and Matterhorn's ingest directory is '/srv/matterhorn/ingest' and mounted on the Avalon server at the same location and Avalon's dropbox directory is '/srv/avalon/dropbox' then Settings.matterhorn.media_path should be configured to be '/srv/matterhorn/ingest'.  Note that Matterhorn's ingset directory needs to be mounted at the same location on both the Avalon and Matterhorn servers for Avalon to be able to ingest due to the use of absolute paths and not relative paths.

Co-authored-by: phuongdh <phuongdh@gmail.com>