Force encoding of IndexedFile original_name to UTF-8
For some reason when a filename with UTF-8 characters is uploaded as a collection poster image
it roundtrips from Fedora as ASCII-8BIT probably due to the fact that it is returned in an
HTTP header.  Forcing the original_name to UTF-8 in an override feels safe after reviewing Avalon's
use of IndexedFiles.

Co-authored-by: Mason Ballengee <masaball@iu.edu>