Do not include sourceless other_identifiers in proxy
The solr_doc includes the master file ids in the other identifier field.
These ids are not included in the accessor on the media object. These
ids also are of a different format than the media object's normal
other_identifiers. We need to exclude the master file ids from the
accessor on the proxy because of those factors.