Reduce ffmpeg timeout to 1 minute
In a manual test using the same ffmpeg commands outside of ruby and without the timeout it took 7 seconds to run for a 1 hour audio file and 25 seconds for a 2 hour video file.  Given this I think a 1 minute timeout should be enough and keeps the queue from getting clogged as much when the stream doesn't close nicely for ffmpeg.