Turn off feeding caption tracks into embedded player
Captions are currently contained in the HLS manifest and these are
parsed by VideoJS. We do not need to manually provide text tracks to the
embedded player because of this. This commit also includes a change to
the filtering of captions for the embedded player for if/when we reenable
manually providing them so that issues will not arise for implementers who
run the caption file migration.