Serialize captions to IIIF manifest with text/vtt as the format
SRT formatted captions are converted to VTT format when the content is
requested by Ramp. However, the iiif manifest still had the format for
these files listed as 'text/srt'. This caused Ramp to not process the
file correctly and the caption file would not show up in the
subs/captions list in the player. Setting the format to 'text/vtt'
allows Ramp to correctly process the caption track.