Ensure that captions as transcript format values are correct
Captions are streamed out as VTT files so the format in the iiif
manifest must be 'text/vtt'. Transcripts do not have this limitation and
so an SRT caption file being treated as a transcript will retain the SRT
formatting. In this case, the iiif manifest must have 'text/srt' as the
format for the transcript annotation.