Add option to rescue Solr/Fedora connection errors
This commit uses a static error page for the Solr connection errors.
The nav bar header rendered as part of the regular Avalon layout checks
user permissions, which directly calls out to Solr. This triggers an
error that is not caught by rescue methods in application_controller.
Because the nav bar is rendered on almost every page of the application,
this results in most of the application being inaccessible. As such,
serving a static error page/message should be sufficient.

Using the Rails `config.action_dispatch.rescue_response` handling
pathway was not logging the proper error in testing and did not prevent
other partials from attempting to render which resulted in extra errors
and noise being present in the logs. Serving this file without the
Avalon layout and logging the Solr error as part of a standard
`rescue_from` in the application_controller feels more in line with our
needs.