Improve handling of timecues in transcript parser
Malformed time cues would cause the transcript search to bomb because
the regex to separate timed text cues would fail to match. We are
loosening the regex to match timecues of the form `h:mm:ss.ttt` and also
providing a failure path so that if the timed text cannot be separated
properly, the search will still return matches. It will just be the
entire cue instead of just the text with a time fragment anchored
target.