Fix playlist description display issues
This commit limits the height of the description section so that it does
not overwhelm the playlist items section and does not extend past the
bottom of the page. With limiting the height we also enable scrolling of
the description when it overflows and mimic the structured nav "scroll
to see more" message.