Fail SupplementalFile Binary update if no Accept header
Users updating supplemental files via CLI should use `-H
"Accept:application/json"` in their requests to ensure a JSON response
from an endpoint. For updating the binary file specifically, the request
is through the HTML format endpoint and so having Accept headers needs
to be required to force a JSON response in the CLI environment.