Skip transcript search in bookmark context
The "Selected Items" page was returning a 400 error because the search
service that populates the page runs through our SearchBuilder model and
generates a `q` solr_parameter. This triggered the transcript search
logic. Something about the `q` generated by the search service broke the
query building for `transcript_tsim`, possibly that the search service
generates `q="{!lucene}(#object_ids)"`. Checking if we are in the
bookmark controller context and skipping the transcript searching allows
us to bypass the error.