Fix enabling of media object action buttons
On page load, both initial and refreshed, in mobile browsers there was
some sort of race condition that was disabling the Add to Playlist,
Timeline, and Thumbnail buttons immediately after being enabled.
Changing the event listener for disabling the buttons to be triggered on
the player emptying instead of loading seems to fix that race condition,
but uncovered a separate one when switching canvases. This second race
condition behaved similarly to the first, where the enabling function
would be run before the disabling function. Adding a fallback enabling
function that triggers on loadstart seems to fix this second race
condition. There may still be some flakiness, especially on apple
devices, because events triggering properly sometimes just doesn't
happen.