Fix broken tests, skip incomplete tests
There was a failing spec because rspec has been updated to raise an
ArgumentError when an `include` matcher is empty. I am temporarily
skipping those tests as I am not familiar with that bit of code so don't
know what is needed to fill them out and get them passing.