Adjust non-media item filtering in ffprobe
In QA it was found that a video was improperly processed as audio. This
was a result of keying off display aspect ratio to filter out
non-audio/video items, under the false assumption that all videos would
have a display aspect ratio. Explicitly filtering out plain text files
and using frame count to filter images will hopefully be more
consistent.