Fix processing of google drive files
When we made the switch from mediainfo to ffprobe, we created a new
class for processing the file. Because Google Drive utilizes
headers/params for its authorization flow, this change resulted in the
file information becoming divorced from the auth token. By explicitly
passing the auth token through to the classes downloading and processing
the file, we are able to fix this oversight.