Ensure activeTrack is set for add current track to playlist save action Related issue: #6235
On MacOS devices the "Add current track" option in the add to playlist
menu was not successfully adding the track in certain circumstances.
This arose from `activeTrack` being undefined in the listener on the
save button. It is still unclear why this value was not available as
expected on MacOS and was fine on other desktop environments, but adding
a fallback to set the `activeTrack` when it is undefined resolves the
issue.