MediaObject#destroy should not fail when it can't find all sections
Sometimes sections get deleted but the parent media object doesn't get updated to remove the
section's id from section_ids.  In this scenario avalon will currently fail to delete the
remaining sections and the media object itself.  This PR adds a safe_load mode to #sections which
only loads sections that are in both solr and fedora.  This mode is used to ensure that
MediaObject#destroy will not fail when it is in an inconsistent state and some of its sections
have already been deleted.