Update popovers to new syntax
An implementation detail is that this leverages Rails confirmations.
https://guides.rubyonrails.org/v6.0/working_with_javascript_in_rails.html#confirmations
The anchor element is initialized so that Rails will generate a
confirmation prompt which we preempt by having an event listener cancel
it and then we render our own confirmation popover. This is a workaround
for an issue where the popover was not automatically preventing the link
navigation and trying to block the action in a click listener was not
working.

This commit also cleans up some old code related to popovers that did
not appear to be getting used.