Allow configuring streaming auth callback referer header
When setting up Wowza in a local dev environment I needed to pass a referer other than rails root in order for wowza to contact avalon.  This could also be useful for other infrastructure setups where the streaming server could make auth calls to avalon on an internal address instead of public address.