Fix date helper to handle "unknown/unknown"
Related issue: #6430

In the original implementation of the view helper for human readable
dates an assumption was made about "unknown/unknown" being edtf
compliant. This was not the case, so the edtf-humanize gem errored in an
environment containing that value. Adding a guard clause to handle that
case takes care of the issue.