Re-enable confirmation messages on delete buttons
Related issue: #6444

Bootstrap 5 changed how popovers are enabled and triggered for display.
As part of the rework for our confirmation popovers, we changed them
from triggering on click to triggering on a rails confirmation event.
During the upgrade some of the delete buttons were missed and so the
rails confirmation was not added to those buttons.