Make edtf humanization configurable
Related issue: #6463

An issue was run into where we fixed date handling for the blacklight
helper but failed to apply it to the iiif manifest generation as well.
This commit moves the humanization into a lambda function in the avalon
configuration. This allows a one stop place to adjust the logic and will
allow other implementers to adjust the humanization to their
specifications.