Update "ffmpeg" location in "config/settings.yml", replace overrides
Since at least Avalon v6.0, the “ffmpeg” binary has been in the
“/usr/bin/” directory, however the "config/settings.yml" file still
sets the path to "/usr/local/bin/ffmpeg".

Assuming this was just an historical oversight, updated the
"config/settings.yml" with the correct location, and removed the
"SETTINGS__FFMPEG__PATH" overrides in the "docker-compose.yml"
and "podman-compose.yml" files, as they should no longer be needed.

This change simplifies the configuration settings, and makes the stock
"settings.yml" file more correct, and usable in more situations (such
as when running Avalon in Kubernetes where the docker-compose.yml"
and "podman-compose.yml" files are not used).