Move event listeners out of apply_button_confirmation function
Related issue: #6493

The event listeners for our confirmation popovers were being initialized
within the `apply_button_confirmation` function. This caused the
listeners to be added multiple times on pages that had dynamic page
reloading, such as the datatables on the playlist and timeline pages.
Moving the listeners outside of the function initializes them only once,
fixing the issue.