Provide more user friendly errors when playlist/timeline is missing
Related issue: #6527

This commit adds a rescue to the playlist and timelines controllers in
order to provide better error pages for when a user tries to access a
deleted or otherwise missing playlist/timeline. For simplicity, this
work reuses the 'errors/unknown_pid' partial but it may eventually
be worthwhile to differentiate between the different contexts.