Override `Autocomplete.open` instead of `Autocomplete.onInputBlur`
The original implementation inadvertently relied on a browser event
property that is only present on firefox. After looking at the original
source code in the library and some more experimentation, it looks like
we can override a different function to set the library's
`interactingWithList` variable to false when the dropdown is initially
opened. We can then rely on the `onResultsMouseDown` function to switch
that variable back to true when a result is clicked on and trigger the
proper selection behavior.