Transition Avalon to jsbundling-rails
It is important to note that the react-on-rails gem is developed by the
developers of shakapacker and is expected to be used in conjunction with
it. Everything seems to be loading and working as I would expect but
there may eventually be updates to react-on-rails or advanced features
we eventually want to use which could break the integration. It seems an
acceptable risk to me since we can pin to a specific version if needed,
barring security issues.