Use cssbundling for delivering most of our stylesheets
There is an incompatibility with the modern scss transpilers where
font-awesome-rails is using a .css.erb file as its main stylesheet, but
the newest transpilation gems and the node based transpilers cannot
parse it. So for the moment this work still uses sassc-rails solely for
bringing in the font-awesome styles. There may be a workaround that
would allow us to move off sassc-rails but I have not been able to get
anything to work. As long as we are on sprockets though, sassc-rails
will be required because it is a dependency of sprockets. If we can find
a way to bring in font-awesome without sassc-rails, we could just have
it sitting in our dependency tree but not actually using it. This is
also not ideal.

Looking forward toward propshaft there are some hurdles that we need to
overcome, primarily how we handle our static javascript files. There is
a comment with full details on issue #6570.