Add Units scaffold and initial code
This commit creates (hopefully) all the controllers, models, and views
for the new Units structure. Initial code for the Units is copied from
the Admin::Collection model and given some tweaking. Tests are not
passing in the work for this commit.

TODO: Properly map hierarchy of permissions for Unit. Do managers,
editors, depositors have permissions to do anything to the Unit or are
they only relevant for being assigned to child collections? If they do
have permissions, how do they differ from the new unit administrator
role?