Add belongs_to association to Admin::Collection
This commit reworks collections to belong_to a Unit. Is this the
approach that should be taken? This would require a large effort data
remediation to create Unit entries for all existing units in a
repository and making sure that all collections are assigned to the
proper new Unit, while retaining all the current permissions structures.
Is there a way to have both the modelled Units and the controlled vocab
based units exist alongside each other to ease the remediation burden?