Use SpeedyAF to scan collection dropboxes at scale
MCO has 800+ collections so loading each one from Fedora takes a while (~4.5m) versus proxies from solr (<0.5s).  This job is scheduled to run every minute so we want it to be fairly fast so catalogers who upload batches don't have to wait 5 minutes to see if they have an error in their manifest or not.  As well as reduce load on Fedora and sidekiq.