Use autocomplete for unit selection on collection new/edit
This commit makes use of the auto-complete-element
'data-autocomplete-value' feature which allows a separate input value to
be assigned to a given display value. The issue is that when a selection
is made, the input is changed to the input value. So a user inputting
'Defaul' then selecting 'Default Unit' will see the input field change
to the unit id. This could be confusing for users, but might be alright?