Prevent m3u8 reload from being added multiple times
Related issue: #6635

The root cause of the behavior was a race condition where the first load
would sometimes take long enough for the m3u8 reload initialization to
be triggered twice or more. This resulted in multiple event listeners,
each creating a reload interval when pausing playback. However, these
intervals were being set to the same variable so when playback was
resumed the interval removal would only remove the most recently created
interval. Using a flag specific to the reload being added to the page
should fix the behavior.