Fix browse everything upload
In the rework to remove jquery the "uploading" modal generated when
submitting a browse everything upload was accidentally broken. We were
attempting to set two `const` variables to the same name, `uploadModal`.
This triggered a reference error which would crash the upload process.
Setting the two variables to different names resolves the issue.