Only request magic bytes for mimetype verification
Related issue: #6649

We analyze input files on upload to determine if they are audio or
video and to grab some of the metadata to fill out the record. This
process involved checking the file mimetype, which was using our
`FileLocator.reader` method. In S3 environments this would download the
entire file, resulting in linearly increasing upload times based on
file size. By requesting just the first ~16 bytes of the file for this
case, we should be able to reduce upload time of multi-gig files from
several minutes down to less than 10 seconds.